improved volume setting for dfplayer mini

+++ src/clients/awb_esp32_client/src/Actuators/Mp3Player/GlobalMp3PlayerManager.h
@@ -24,7 +24,7 @@ public:
 
     bool playSound(String playerTitle, int trackNo);
     bool stopSound(String playerTitle);
-    bool setVolume(String playerTitle, int volume);
+    bool setVolume(String playerTitle, uint8_t volume);
 };
 
 #endif
