+++ src/clients/awb_esp32_client/src/Actuators/Mp3Player/Mp3PlayerDfPlayerMiniManager.cpp
@@ -22,7 +22,7 @@ bool Mp3PlayerDfPlayerMiniManager::stopSound(int playerIndex)
     return false;
 }
 
-bool Mp3PlayerDfPlayerMiniManager::setVolume(int playerIndex, int volume)
+bool Mp3PlayerDfPlayerMiniManager::setVolume(int playerIndex, uint8_t volume)
 {
     if (playerIndex < _mp3Players->size())
     {
